requirements.txt
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-User==1.0.2.2
Flask-WTF==1.0.0
pandas==1.3.5
Flask-Cors==3.0.10
influxdb
networkx
junos-eznc
numpy
nornir>=3.0.0
nornir_napalm
nornir_netmiko
nornir_utils
nornir-jinja2
argon2_cffi
simple-crypt
pycryptodome==3.5
pynetbox
redis
cryptography
flask_httpauth
genie
pyats
slackclient
pymsteams
rich
